+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Notes for Multi Translator
 
+## 2.28.0 - 2026-06-09
+
+### Added
+
+- refactor api services to providers and allow for custom API provider registration
+
 ## 2.27.1 - 2026-05-29
 
 ### Fixed
@@ -10,7 +16,7 @@
 
 ### Added
 
-- support OpenAI-compatible endpoints (base URL and custom model)
+- support OpenAI-compatible endpoints (base URL and custom model) (thanks @riccardolardi)
 
 ### Fixed